chore: updated README.md
I have continued working on the server and I needed to update the changes I have been making in the README.md file.  More similar updates will come because the server is not finished yet and I will have to update the readme to show how I got it to work.